fix: quick+ to compile on macOS
This fixes an issue that prevented the quick+ artifact from
compiling on macOS. It still compiles fine on Linux. Didn't check
if it compiles on Windows, but I believe it should.

Note there are still some warnings.

PS: Note that on macOS, g++ points to the default compiler (clang)
which fails to compile the quick+ artifact. My solution was to
change the Makefile to use g++-14 (installed via homebrew).